principalEmzima/pythonAutomatizacion/saucedemoprojectPlaywrightAsync/requirements.txt
allure-pytest==2.13.5
behave==1.2.6
pytest==8.3.3
pytest-html==4.1.1
pytest-bdd==8.1.0
playwright==1.49.1
python==3.12.6
pytest-asyncio-0.25.0